Move `ParityDbStore` into farmer's utils, so it is reusable by SDK
